codeql: close java/implicit-cast-in-compound-assignment alerts #423-#429 (#51)
7 high-severity CodeQL alerts closed with explicit narrowing casts and
field-type widening; no behavior change at the unit level:

- system/.../HTTPClient/BufferedInputStream.java:115 — pos = (int)(pos + n)
- system/.../HTTPClient/RespInputStream.java:140 — offset = (int)(offset + num)
- system/.../server/cache/PSAutotuneCache.java:199,216 — accumulator widened
  long → double with final (long) cast (calcSpaceFor{Large,Smaller}Tables)
- system/Testing/.../HttpRequest.java:406,649 — m_bytesSent widened int →
  long; getBytesSent returns (int) m_bytesSent (existing public signature)
- deliverytiersuite/.../PSFeedServicePerformanceTest.java:51 — sum widened
  int → long so sum += longValue no longer truncates (test is @Ignore'd but
  CodeQL still flagged the sink)

clusters.md row updated to point at this PR.

JDK 1.8.0 compatible. Verified with:
- mvn compile -pl system -am
- mvn test-compile -pl deliverytiersuite/delivery-tier-suite/feeds -am
- mvn test -pl system -am -Dtest=PSAutotuneCacheTest

+++ deliverytiersuite/delivery-tier-suite/feeds/src/test/java/com/percussion/delivery/feeds/PSFeedServicePerformanceTest.java
@@ -48,7 +48,7 @@ import org.junit.runner.RunWith;
 @Server(url = "http://localhost:9980/", siteName = "Site1", feedName = "feeds1")
 public class PSFeedServicePerformanceTest {
   private static final Logger log = LogManager.getLogger(PSFeedServicePerformanceTest.class);
-  private static int sum = 0;
+  private static long sum = 0;
   private static int Threshold = 0;
   private static String URL = "";
   private static String SITENAME = "";
